+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 Version numbers correspond to `bower.json` version
 
+# 1.2.2
+
+## Bug Fixes
+
+- Fix redirect loop
+
+# 1.2.1
+
+## Bug Fixes
+
+- Fix issue with login without credentials
+
+# 1.2.0
+
+## Features
+
+- Implement login with redirect uri
+
 # 1.0.0
 
 ## Features